patterns: factory-method: Added sample of a factory method pattern
Added example of use of factory method pattern. The example shows
how thans to facotry method an application may be decupled from
an implementation of an interface. The factory method delegates
instantiation process to an implementation of the method.
An application uses an interface, so it does not know what is
actual type of the object it has access to.

The implementation of the pattern may have couple of features:
1. The virtual method may have a default implementation (as in sample)
that is overriden by derived classes.
2. The factory methot may return an error to inform the instantiation
process failed.
3. The factory method may be a tempate to avoid crating inheritance
tree for creator classes.
4. If the method is not going to be derived it may be turned into
static. Then there is no need for instance of a creator class.
5. Factory method may take arguments that will be used to select
appropriate type of a class to be instantiated.

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>