cpp-constructors: Add class instantiation by static method
Add another aproach to a class instantiation with use of a static
method. Thanks to that one can return a value that indicated
success or failuer of an instance creation.
Example uses std::optional. It allows to store a class instance
by value, hence to heap is used for an object creation.
It also allows to store std::nullopt to indicate error.

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>